% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{calc_CI}
\alias{calc_CI}
\title{Calculate combination index (CI)}
\usage{
calc_CI(drug1, drug2, drug_combo, fa = double())
}
\arguments{
\item{drug1}{Drug effect object}

\item{drug2}{Drug effect object}

\item{drug_combo}{Drug effect fixed-ratio combination object}

\item{fa}{Vector of fraction affected (fa) at which combination indices (CI) will be calculated (optional)}
}
\description{
Given drug 1, drug 2, and drug_combo objects, calculate combination index CI at vector of fa values.
If fa is empty, then use the actual fa and actual doses in drug_combo.
}
