% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{calc_combo}
\alias{calc_combo}
\title{Drug combination calculations}
\usage{
calc_combo(drug_combo, ..., fa = double())
}
\arguments{
\item{drug_combo}{Drug effect fixed-ratio combination object}

\item{...}{Drug effect objects}

\item{fa}{Vector of fraction affected (fa) at which calculations will be made (optional)}
}
\description{
Given drug combination object (either fixed or non-fixed ratio) and arbitrary number
of single-drug effect objects, calculate
all parameters needed for later calculation of combination or dose reduction index,
at either a vector of specified fa values or using actual doses / fa from drug_combo.
Returns unique id, total dose in combination, fa, single drug doses needed for fa,
and dose in the combination for each drug.
}
