% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{calc_d}
\alias{calc_d}
\title{Calculate predicted doses}
\usage{
calc_d(drug, fa)
}
\arguments{
\item{drug}{Drug effect object}

\item{fa}{Vector of fraction affected (fa)}
}
\description{
Given a drug_effects object and vector of fraction affected (fa), return predicted doses required
}
