% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{calc_DRI}
\alias{calc_DRI}
\title{Calculate dose reduction index (DRI)}
\usage{
calc_DRI(drug_combo, ..., fa = double())
}
\arguments{
\item{drug_combo}{Drug effect fixed-ratio combination object}

\item{...}{Drug effect objects}

\item{fa}{Vector of fraction affected (fa) at which calculations will be made (optional)}
}
\description{
Given drug_combo and arbitrary number of single-drug effect objects, calculate
the dose reduction index for each drug at a vector of fa values. If fa is empty,
then use the actual fa and actual doses in drug_combo.
}
