% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{calc_fa}
\alias{calc_fa}
\title{Calculate predicted fraction affected (fa)}
\usage{
calc_fa(drug, D)
}
\arguments{
\item{drug}{Drug effect object}

\item{D}{Vector of doses}
}
\description{
Given a drug_effects object and vector of doses, return predicted fraction affected
}
