% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{dose_effect_plot}
\alias{dose_effect_plot}
\title{Generate Dose-Effect Plot}
\usage{
dose_effect_plot(..., from = 0.01, to = 0.99, by = 0.01)
}
\arguments{
\item{...}{drug_effects objects to plot}

\item{from}{Start of range of fraction affected (fa)}

\item{to}{End of range of fa}

\item{by}{Step size of fa}
}
\description{
Generates Dose-Effect plots (as ggplot objects) for an arbitrary number of drug_effects objects
}
\details{
Plots curves of fa versus D.
}
