% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{drug_combo_decompose}
\alias{drug_combo_decompose}
\title{Function to decompose a drug_combo drug effects object into its component drug doses}
\usage{
drug_combo_decompose(drug_combo)
}
\arguments{
\item{drug_combo}{A fixed-ratio combination drug-effects object}
}
\description{
Function to decompose a drug_combo drug effects object into its component drug doses
}
