% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{drug_effects}
\alias{drug_effects}
\title{Drug object generator}
\usage{
drug_effects(
  D = double(),
  fa = double(),
  name = "",
  label = "",
  info = "",
  ratio = double()
)
}
\arguments{
\item{D}{Vector of doses (single drug dose or sum of doses for fixed-ratio combinations)}

\item{fa}{Vector of fraction affected}

\item{name}{Long name of drug (optional)}

\item{label}{Short label of drug (for plot labels, optional)}

\item{info}{Longer description (optional)}

\item{ratio}{If present, specifies vector of ratio of each drug in constant ratio combinations}
}
\description{
Function to create a drug_effects object, either for a single drug or fixed-ratio combination of 2 drugs
}
\examples{
D  <- c(0.10, 0.15, 0.20, 0.25, 0.35)
fa <- c(0.24, 0.44, 0.63, 0.81, 0.90)
drug_effects(D = D, fa = fa, name = 'Example', label = 'drug1')

}
