% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{fa_ci_plot}
\alias{fa_ci_plot}
\title{Generate fa-CI Plot}
\usage{
fa_ci_plot(drug1, drug2, drug_combo, from = 0.01, to = 0.99, by = 0.01)
}
\arguments{
\item{drug1}{Drug effect object}

\item{drug2}{Drug effect object}

\item{drug_combo}{Drug effect fixed-ratio combination object}

\item{from}{Start of range of fraction affected (fa)}

\item{to}{End of range of fa}

\item{by}{Step size of fa}
}
\description{
Generates fa-CI plots (as ggplot objects) for a two-drug fixed-ratio combination
}
