% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{fa_dri_plot}
\alias{fa_dri_plot}
\title{Generate fa-DRI Plot}
\usage{
fa_dri_plot(drug_combo, ..., from = 0.01, to = 0.99, by = 0.01)
}
\arguments{
\item{drug_combo}{Drug effect fixed-ratio combination object}

\item{...}{Drug effect objects}

\item{from}{Start of range of fraction affected (fa)}

\item{to}{End of range of fa}

\item{by}{Step size of fa}
}
\description{
Generates fa-DRI plots (as ggplot objects) for a fixed-ratio combination
}
