% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{isobologram_plot}
\alias{isobologram_plot}
\title{Generate isobologram plot}
\usage{
isobologram_plot(drug_combo, drug1, drug2, normalized = TRUE)
}
\arguments{
\item{drug_combo}{Two drug combination drug effect object}

\item{drug1}{Drug effect object to plot on X axis}

\item{drug2}{Drug effect object to plot on Y axis}

\item{normalized}{Whether to plot normalized (default) or classic isobologram}
}
\description{
Generate either normalized or classic isobologram plot for either fixed ratio or
non-constant ratio two drug combinations. The observed fraction affected (fa) values
from the drug combination will be used
}
