% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{median_effect_plot}
\alias{median_effect_plot}
\title{Generate Median-Effect Plot}
\usage{
median_effect_plot(...)
}
\arguments{
\item{...}{drug_effects objects to plot}
}
\description{
Generates Median-Effect plots (as ggplot objects) for an arbitrary number of drug_effects objects
}
\details{
Plots linearized log(fa / (1-fa)) versus log(D)
}
