% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{ncr_calc_ci}
\alias{ncr_calc_ci}
\title{Calculate combination index (CI) for non-constant ratio}
\usage{
ncr_calc_ci(ncr_combo, ...)
}
\arguments{
\item{ncr_combo}{Non-constant ratio combination drug effects object}

\item{...}{Drug effect objects}
}
\description{
Given non-constant ratio drug effects object and arbitrary number of single-drug
effect objects, calculate combination index observed fa values.
}
