% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{ncr_calc_combo}
\alias{ncr_calc_combo}
\title{Non-constant ratio drug combination calculations}
\usage{
ncr_calc_combo(ncr_combo, ...)
}
\arguments{
\item{ncr_combo}{Non-constant ratio combination drug effects object}

\item{...}{Drug effect objects}
}
\description{
Given non-constant ratio drug effects object and an arbitrary number of single-drug effect
objects, calculate all parameters needed for later calculation of combination index.
}
