% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{ncr_drug_effects}
\alias{ncr_drug_effects}
\title{Non-constant ratio drug effects object generator}
\usage{
ncr_drug_effects(..., fa = double(), name = "", label = "", info = "")
}
\arguments{
\item{...}{Vectors of doses, one for each drug, with length equal to number of data points}

\item{fa}{Vector of fraction affected}

\item{name}{Name to describe set (optional)}

\item{label}{Short label (optional)}

\item{info}{Longer description (optional)}
}
\description{
Function to generate a non-constant ratio drug_effects object
}
