% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{new_drug_effects}
\alias{new_drug_effects}
\title{Drug effects object constructor}
\usage{
new_drug_effects(
  D = double(),
  fa = double(),
  name = "",
  label = "",
  info = "",
  ratio = double()
)
}
\arguments{
\item{D}{Vector of doses (single drug dose or sum of doses for fixed-ratio combinations)}

\item{fa}{Vector of fraction affected}

\item{name}{Long name of drug (optional)}

\item{label}{Short label of drug (for plot labels, optional)}

\item{info}{Longer description (optional)}

\item{ratio}{If present, specifies vector of ratio of each drug in constant ratio combinations}
}
\description{
Internal constructor function to generate a drug_effects object, for single drugs and constant ratio combinations.
}
