% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{plot.drug_effects}
\alias{plot.drug_effects}
\title{Plot method for objects of drug_effects class}
\usage{
\method{plot}{drug_effects}(x, y, ..., color = "blue")
}
\arguments{
\item{x}{A drug_effects object}

\item{y}{(not used)}

\item{...}{(not used)}

\item{color}{Line color}
}
\description{
Plot method for objects of drug_effects class
}
