% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{print.drug_effects}
\alias{print.drug_effects}
\title{Print method for objects of drug_effects class}
\usage{
\method{print}{drug_effects}(x, ..., stats = TRUE)
}
\arguments{
\item{x}{A drug_effects object}

\item{...}{(not used)}

\item{stats}{Boolean for whether to display calculated m, Dm, R^2, and R}
}
\description{
Print method for objects of drug_effects class
}
