% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{print.ncr_drug_effects}
\alias{print.ncr_drug_effects}
\title{Print method for non-constant ratio drug_effects class}
\usage{
\method{print}{ncr_drug_effects}(x, ...)
}
\arguments{
\item{x}{A non-constant ratio drug_effects object}

\item{...}{(not used)}
}
\description{
Print method for non-constant ratio drug_effects class
}
