% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{validate_drug_effects}
\alias{validate_drug_effects}
\title{Drug effects object validator + calculator}
\usage{
validate_drug_effects(x)
}
\arguments{
\item{x}{drug_effects object}
}
\description{
Internal validator + calculator function.
}
\details{
Checks that doses are non-negative, fraction affected is between (0,1),
equal length vector of at least two doses and effects, ratio positive (if present).

Also calculates log(D) and log(fa / (1-fa)), fits a linear model, and calculates b, m, Dm (median effect dose), and
R^2 statistic (as well as R).

Returns the drug_effects object augmented with the additional calculated information.
}
