% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianeffect.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculate combination index (CI)}
\usage{
calc_ci(drug_combo, ..., fa = double())
}
\arguments{
\item{drug_combo}{Drug effect fixed-ratio combination object}

\item{...}{Drug effect objects}

\item{fa}{Vector of fraction affected (fa) at which calculations will be made (optional)}
}
\description{
Given drug_combo and arbitrary number of single-drug effect objects, calculate
combination index at vector of fa values. If fa is empty, then use the actual fa
and actual doses in drug_combo.
}
